requirements.txt
gymnasium==0.29.1
stable-baselines3==2.2.1
torch==2.1.2
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0.1
scikit-learn==1.3.2
matplotlib==3.7.2
ta==0.10.2  # Technical analysis library
tensorboard==2.14.1  # For training visualization 